Textures excersises

+++ src/shader.h
@@ -106,7 +106,6 @@ public:
         glUniform1f(vertexHorizontalOffsetLocation, x);
     }
 
-
     // utility uniform functions
     void setBool(const std::string &name, bool value) const
     {
@@ -114,11 +113,11 @@ public:
     }
     void setInt(const std::string &name, int value) const
     {
-        glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+        glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
     }
     void setFloat(const std::string &name, float value) const
     {
-        glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+        glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
     }
 };
